app/api/v1/schemas/conversation_schemas.py
from datetime import datetime
from uuid import UUID

from pydantic import BaseModel, ConfigDict

from app.domain.entities import MessageRole


class ConversationCreateRequest(BaseModel):
    customer_id: UUID
    title: str


class ConversationResponse(BaseModel):
    id: UUID
    customer_id: UUID
    title: str
    created_at: datetime
    updated_at: datetime

    model_config = ConfigDict(from_attributes=True)


class MessageSendRequest(BaseModel):
    content: str


class MessageResponse(BaseModel):
    id: UUID
    conversation_id: UUID
    role: MessageRole
    content: str
    created_at: datetime

    model_config = ConfigDict(from_attributes=True)
